+++ packages/adhere-ui-conditionalrender/src/conditionalrender.tsx
@@ -22,6 +22,44 @@ function ConditionalRender({ conditional, noMatch, children }: IConditionalRende
 ConditionalRender.Show = ConditionalRenderShow;
 ConditionalRender.Visibility = ConditionalRenderVisibility;
 
+/**
+ * conditionalRender
+ * @description - 使用方法的ConditionalRender
+ * @param conditional
+ * @param match
+ * @param noMatch
+ */
+ConditionalRender.conditionalRender = function ({
+  conditional,
+  match,
+  noMatch,
+}: {
+  conditional: boolean;
+  match: JSX.Element;
+  noMatch: JSX.Element | null;
+}): JSX.Element | null {
+  return conditional ? match : noMatch || null;
+};
+
+/**
+ * conditionalArr
+ * @description - 含有PermissionConditional的React.Element的数组
+ * @param arr
+ * @return Array
+ */
+ConditionalRender.conditionalArr = function (arr: any[]): any[] {
+  return arr.filter((t) => {
+    if ('props' in t && 'conditional' in t.props) {
+      if (!t.props.conditional) {
+        if (t.props.noMatch && t.props.noMatch?.() !== null) return true;
+        return false;
+      }
+    }
+
+    return true;
+  });
+};
+
 ConditionalRender.defaultProps = {
   conditional: true,
   noMatch: () => null,